+++ markwon-inline-parser/src/main/java/io/noties/markwon/inlineparser/BangInlineProcessor.java
@@ -29,10 +29,6 @@ public class BangInlineProcessor extends InlineProcessor {
 
             return node;
         } else {
-            // @since 4.5.0 return null in case no match (multiple inline
-            //  processors can define `!` as _special_ character, so let them handle it)
-            //  NB! do not forget to reset index
-            index = startIndex;
             return null;
         }
     }