+++ markwon-inline-parser/src/main/java/io/noties/markwon/inlineparser/MarkwonInlineParser.java
@@ -266,11 +266,18 @@ public class MarkwonInlineParser implements InlineParser, MarkwonInlineParserCon
         final List<InlineProcessor> inlines = this.inlineProcessors.get(c);
 
         if (inlines != null) {
+            // @since $SNAPSHOT; index must not be advanced if inline-processor returned null
+            //  so, further processors can be called at the _same_ position (and thus char)
+            final int startIndex = index;
+
             for (InlineProcessor inline : inlines) {
                 node = inline.parse(this);
                 if (node != null) {
                     break;
                 }
+
+                // reset after each iteration (happens only when node is null)
+                index = startIndex;
             }
         } else {
             final DelimiterProcessor delimiterProcessor = delimiterProcessors.get(c);